Ajuste de la estructura
Se realiza la creacion de tablas
se ajusta la base de la estructura del codigo en netbens

+++ src/main/java/com/concentrados/Danny/controller/RegistroController.java
@@ -35,7 +35,7 @@ public class RegistroController {
 
     // 3. Procesar el cambio de rol desde la tabla de gestión
     @PostMapping("/asignarRol")
-    public String asignarRol(@RequestParam("idUsuario") Integer idUsuario, 
+    public String asignarRol(@RequestParam("idUsuario") Long idUsuario, 
                              @RequestParam("nombreRol") String nombreRol) {
         
         usuarioService.asignarRol(idUsuario, nombreRol);
